add support for shifted option for variable collation elements
... i.e. algorithm step
http://www.unicode.org/reports/tr10/tr10-36.html#S2.3

In order to achieve that, this commit extracts info on variable
collation elements from allkeys data file as specified at
https://www.unicode.org/reports/tr10/tr10-36.html#File_Format

Description of conformance test is at
https://www.unicode.org/Public/UCA/10.0.0/CollationTest.html

Sample comparison of non-ignorable vs. shifted ordering is at
http://www.unicode.org/reports/tr10/tr10-36.html#Variable_Weighting_Examples

Intended meaning of combination of variable weighting (e.g. shifted)
and strength is described in the following locations:
* https://www.unicode.org/reports/tr10/tr10-36.html#Parametic_Tailoring
* https://www.unicode.org/reports/tr35/tr35-collation.html#Common_Settings